+++ modules/example-browser/karma.conf.js
@@ -1,3 +1,6 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 const webpack = require('webpack')
 const {defaultProvider} = require('@aws-sdk/credential-provider-node')