+++ modules/example-node/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/aws/aws-encryption-sdk-javascript/compare/@aws-crypto/example-node@1.0.4...@aws-crypto/example-node@1.0.5) (2020-05-26)
+
+**Note:** Version bump only for package @aws-crypto/example-node
+
+
+
+
+
 ## [1.0.4](https://github.com/aws/aws-encryption-sdk-javascript/compare/@aws-crypto/example-node@1.0.3...@aws-crypto/example-node@1.0.4) (2020-04-02)
 
 **Note:** Version bump only for package @aws-crypto/example-node