+++ modules/hkdf-node/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/aws/aws-encryption-sdk-javascript/compare/@aws-crypto/hkdf-node@1.0.2...@aws-crypto/hkdf-node@1.0.3) (2020-05-26)
+
+**Note:** Version bump only for package @aws-crypto/hkdf-node
+
+
+
+
+
 ## [1.0.2](https://github.com/aws/aws-encryption-sdk-javascript/compare/@aws-crypto/hkdf-node@1.0.1...@aws-crypto/hkdf-node@1.0.2) (2020-04-02)
 
 **Note:** Version bump only for package @aws-crypto/hkdf-node