+++ modules/web-crypto-backend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/aws/aws-encryption-sdk-javascript/compare/@aws-crypto/web-crypto-backend@1.1.1...@aws-crypto/web-crypto-backend@1.1.2) (2020-05-26)
+
+**Note:** Version bump only for package @aws-crypto/web-crypto-backend
+
+
+
+
+
 ## [1.1.1](https://github.com/aws/aws-encryption-sdk-javascript/compare/@aws-crypto/web-crypto-backend@1.1.0...@aws-crypto/web-crypto-backend@1.1.1) (2020-04-02)
 
 **Note:** Version bump only for package @aws-crypto/web-crypto-backend